requirements.txt
discord.py
python-dotenv
python-dotenv[cli]